split integration tests
