+++ tox.ini
@@ -75,15 +75,20 @@ deps =
     pytest-order
     -r{toxinidir}/requirements.txt
 commands =
-    # Generate TLS certificates to use during tests
-    sh -c "./tests/gen-test-certs.sh"
-    pytest -v --tb native --ignore={[vars]tst_path}unit --log-cli-level=INFO -s {posargs} --durations=0 {[vars]tst_path}integration/test_charm.py
-    sh -c "rm -f tests/tls/*"
-whitelist_externals =
-    sh
+    pytest -vv --tb native --ignore={[vars]tst_path}unit --log-cli-level=INFO -s {posargs} --durations=0
 
-[testenv:integration-relation]
-description = Run integration tests
+[testenv:integration-charm]
+description = Run charm integration tests
+deps =
+    pytest
+    juju
+    pytest-operator
+    -r{toxinidir}/requirements.txt
+commands =
+    pytest {[vars]tst_path}/integration/test_charm.py -vv --tb native --ignore={[vars]tst_path}unit --log-cli-level=INFO -s {posargs} --durations=0
+
+[testenv:integration-scaling]
+description = Run scaling integration tests
 deps =
     pytest
     juju
@@ -91,4 +96,24 @@ deps =
     pytest-order
     -r{toxinidir}/requirements.txt
 commands =
-    pytest -v --tb native --ignore={[vars]tst_path}unit --log-cli-level=INFO -s {posargs} -m "redis_tests" --durations=0
+    pytest {[vars]tst_path}/integration/test_scaling.py -vv --tb native --ignore={[vars]tst_path}unit --log-cli-level=INFO -s {posargs} --durations=0
+
+[testenv:integration-password]
+description = Run password integration tests
+deps =
+    pytest
+    juju
+    pytest-operator
+    -r{toxinidir}/requirements.txt
+commands =
+    pytest {[vars]tst_path}/integration/test_password.py -vv --tb native --ignore={[vars]tst_path}unit --log-cli-level=INFO -s {posargs} --durations=0
+
+[testenv:integration-relation]
+description = Run integration tests for redis relation
+deps =
+    pytest
+    juju
+    pytest-operator
+    -r{toxinidir}/requirements.txt
+commands =
+    pytest {[vars]tst_path}/integration/test_redis_relation.py -vv --tb native --ignore={[vars]tst_path}unit --log-cli-level=INFO -s {posargs} --durations=0
